Major overhaul to the I/O functionality:
The SequentialStremMultiplexer (SSM) has been replaced by AosIOStream.
The AosIOStream class acts like a BaseSequentialStream but consists of multiple AosIOChannels.
Each AosIOChannel again implements a BaseAsynchronousChannel and thus is a derived class of BaseSequentialStream as well.
A similar implementation has been done for the shell (AosShellStream and AosShellChannel).
Please note that the *Stream classes do not support input functionality yet.